code/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.4
django-debug-toolbar==3.2.4
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
numpy==1.22.3
pandas==1.4.2
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0

